requirements.txt
# Interface Web
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# PDF Processing Core
PyPDF2>=3.0.0
pdfplumber>=0.9.0
PyMuPDF>=1.23.0
pikepdf>=8.0.0

# OCR and Optimization
OCRmyPDF>=15.0.0

# Image Processing
Pillow>=10.0.0

# PDF Conversion Dependencies
pdf2docx>=0.5.6
python-docx>=0.8.11
openpyxl>=3.1.0
pypandoc>=1.11
ebooklib>=0.18
lxml>=4.9.0
beautifulsoup4>=4.12.0

# Utilities
pathlib2
psutil

# Legacy GUI (optional - for backward compatibility)
# customtkinter